+++ tests/Classifiers/RandomForestTest.php
@@ -57,7 +57,7 @@ class RandomForestTest extends TestCase
             'blue' => new Blob([0, 32, 255], 20.0),
         ], [2, 3, 4]);
 
-        $this->estimator = new RandomForest(new ClassificationTree(3), 100, 0.2);
+        $this->estimator = new RandomForest(new ClassificationTree(3), 100, 0.2, true);
 
         $this->estimator->setBackend(new Serial());
 
@@ -123,6 +123,7 @@ class RandomForestTest extends TestCase
             'base' => new ClassificationTree(3),
             'estimators' => 100,
             'ratio' => 0.2,
+            'balanced' => true,
         ];
 
         $this->assertEquals($expected, $this->estimator->params());